Between

+++ src/main/java/ru/skypro/homework/dto/AdsDTO.java
@@ -1,10 +1,15 @@
 package ru.skypro.homework.dto;
 
 import lombok.Data;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.List;
 
 @Data
 public class AdsDTO {
-    private Long author;
+    private Integer author;
     private String image;
     private Integer pk;
     private Integer price;
