+++ src/main/java/ru/skypro/homework/dto/CommentDTO.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class CommentDTO {
-    private Long author;
+    private Integer author;
     private String authorImage;
     private String authorFirstName;
     private Integer createdAt;
-    private Long pk;
+    private Integer pk;
     private String text;
 }
