+++ src/main/java/ru/skypro/homework/model/Comment.java
@@ -9,7 +9,7 @@ import javax.persistence.*;
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long pk;
+    private Integer pk;
     private String authorImage;
     private Integer createdAt;
     private String text;
